chore: move PageHeaderSection
