+++ src/pages/challenges/[id]/teams/join.tsx
@@ -1,7 +1,7 @@
 import ChallengeTitle from '@/components/common/challenges/challenge-title'
 import TeamCard from '@/components/common/teams/team-card'
-import PageContainer from '@/components/common/page-container'
-import PageHeaderSection from '@/components/common/page-header-section'
+import PageContainer from '@/components/common/page-layout/container'
+import PageHeaderSection from '@/components/common/page-layout/header-section'
 import PageTitle from '@/components/common/page-title'
 import useChallengesTeams from '@/hooks/challenge/use-challenges-teams'
 import { useNavigate, useParams } from 'react-router-dom'
