+++ src/pages/point-shop/index.tsx
@@ -1,6 +1,6 @@
 import BottomNavigation from '@/components/common/bottom-navigation'
 import PageContainer from '@/components/common/page-layout/container'
-import PageHeaderSection from '@/components/common/page-header-section'
+import PageHeaderSection from '@/components/common/page-layout/header-section'
 import PageTitle from '@/components/common/page-title'
 import ProductList from '@/components/shop-screen/product-list'
 import UserStatusbar from '@/components/shop-screen/user-statusbar'
