+++ src/pages/posts/index.tsx
@@ -5,7 +5,7 @@ import { TabType } from '@/components/post-screen/category-tab/types'
 import { usePostsArrayOnly } from '@/hooks/post/use-posts'
 import { useState } from 'react'
 import PageContainer from '@/components/common/page-layout/container'
-import PageHeaderSection from '@/components/common/page-header-section'
+import PageHeaderSection from '@/components/common/page-layout/header-section'
 import PageTitle from '@/components/common/page-title'
 import { CircleAlert } from 'lucide-react'
 import useIsLoggedIn from '@/hooks/use-is-logged-in'
